first draft at SUPG type of stabilization for flow solver
Seems to work at least up to Re = 3200 on the default Gmsh-imported mesh
at the default timestep size (T = 5, nt = 2500, so dt = 2e-3).

Still need to test against a high-resolution solution. The stabilization
is in any case consistent and seems to help at high Re, but I'm not sure
whether the weighting function is correct; this is adapted from Donea &
Huerta (2003), but slightly different from any of theirs.

For testing, to set Re = 3200, in config.py, divide μ by 16.