allow compiling also with DG(0) pressure
The correction step of IPCS (as implemented here) needs at least P1 elements, so
this won't actually compute anything useful when DG(0) is used. This commit just
fixes a crash that gives a mysterious error message (integral is missing domain)
when attempting to use DG(0).

The issue was that UFL doesn't seem to be able to automatically figure out the
integration domain for (at least) some integrands that are identically zero.