euleriansolid: fix plotting slowdown/eventual hang
Now each plotting update takes roughly the same amount of wall time throughout
the simulation.

The trick is to call `plt.tight_layout()` explicitly, instead of relying on
the `constrained_layout` option of `plt.subplots()`. In the old solution, the
symptom is that each time a colorbar is created, plotting becomes slower,
and after enough colorbars have been created (in total, even though old ones
are removed from the figure; the total limit seems to be ~300), the plotter
hangs upon creating a colorbar.

The explicit layout call causes the subplots to twitch a little at each update
as they might get slightly repositioned, but that's much better than a major
slowdown and a hang.

Also, we now use the OOP API of Matplotlib.

Using the OOP API, one of the subplots (elastic strain energy) required manual
xlim/ylim to display correctly, for some reason.

We now scan for the bounding box upon startup, and then use it as xlim/ylim in
all subplots. This also has the nice effect of avoiding empty space in each axes
around the actual data (so the area inside the axes is used maximally to show
data).