fix EulerianSolidPrimal
Turns out the only thing that was needed was to switch to a direct linear
solver. MUMPS (MUltifrontal Massively Parallel Sparse direct Solver) finds
the correct solution, and makes this solver blazingly fast.

And, well, the boundary condition on `v` may also have something to do with it,
but that alone wasn't enough.