change the ordering of the weak coupling
Since the 3D printing problem is mainly thermally driven, swap the order of the
substeps: solve the thermal subproblem first. This should help especially at the
beginning of a simulation (the first system iteration of the first timestep),
where the initial condition might not be a physically valid temperature field.

Solving the thermal subproblem first effectively smooths out the temperature
field at the end of the timestep, before it is used for computing thermal
strains in the mechanical subproblem.